refactor safe writer
